Put tests in more correct place
